`use-triggered-from` boolean: links upload to triggering build
Uses the `BUILDKITE_TRIGGERED_FROM_…` environment variables to attribute
the uploaded tests to the build that triggered this one.

This is useful when the tests are run in one build, and their results
(e.g. junit.xml) are uploaded by a separate triggered build. By default,
those tests would appear to Test Engine as running within the triggered
build that uploaded them, and so they would not be shown in the UI of
the build that actually ran them.